requirements.txt
flask
mysql-connector-python
python-dotenv
werkzeug
gunicorn
yfinance
faker
pandas
